pyproject.toml
[tool.poetry]
name = "hanyuu"
version = "0.1.0"
description = ""
authors = ["mozhaa <mozhay2005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
aiohttp = "^3.11.7"
sqlalchemy = "^2.0.36"
aiosqlite = "^0.20.0"
greenlet = "^3.1.1"
pydantic-settings = "^2.6.1"
fastapi = {extras = ["standart"], version = "^0.115.5"}
asyncpg = "^0.30.0"
lxml = "^5.3.0"
orjson = "^3.10.12"
rapidfuzz = "^3.10.1"
filelock = "^3.16.1"
aiofiles = "^24.1.0"
pyquery = "^2.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
